Update task 6 requirements
In task 6 the `add()` function should be a const member function
because `this` instance that the add is called on does not, and
should not, be modified as a result of performing the add operation.
A new `LargeInteger` is created to hold the result and returned, but
neither `this` nor the `other` instance should be modified.